+++ FireAxe.GUI/Program.cs
@@ -15,7 +15,7 @@ internal sealed class Program
     // SynchronizationContext-reliant code before AppMain is called: things aren't initialized
     // yet and stuff might break.
     [STAThread]
-    public static void Main(string[] args)
+    public static int Main(string[] args)
     {
         _ = LanguageManager.Instance;
         SetupLogger();
@@ -25,8 +25,7 @@ internal sealed class Program
         {
             RegisterObjectExplanations();
 
-            BuildAvaloniaApp()
-            .StartWithClassicDesktopLifetime(args);
+            return BuildAvaloniaApp().StartWithClassicDesktopLifetime(args);
         }
         catch (Exception ex)
         {
@@ -37,6 +36,7 @@ internal sealed class Program
         finally
         {
             Log.CloseAndFlush();
+            AppMutex.Release();
         }
     }